Public API

GPPP

Core functionality for working with GPPPs.

Stheno.@gppp — Macro

@gppp(model_expression)

Construct a GaussianProcessProbabilisticProgramme (GPPP) from a code snippet.

f = @gppp let
    f1 = GP(SEKernel())
    f2 = GP(Matern52Kernel())
    f3 = f1 + f2
end

x = GPPPInput(:f3, randn(5))

y = rand(f(x, 0.1))

logpdf(f(x, 0.1), y) ≈ elbo(f(x, 0.1), y, f(x, 1e-9))

# output

true

Stheno.GPPPInput — Type

GPPPInput(p, x::AbstractVector)

An collection of inputs for a GPPP. p indicates which process the vector x should be extracted from. The required type of p is determined by the type of the keys in the GPPP indexed.

julia> x = [1.0, 1.5, 0.3];

julia> v = GPPPInput(:a, x)
3-element GPPPInput{Symbol, Float64, Vector{Float64}}:
 (:a, 1.0)
 (:a, 1.5)
 (:a, 0.3)

julia> v isa AbstractVector{Tuple{Symbol, Float64}}
true

julia> v == map(x_ -> (:a, x_), x)
true

Stheno.BlockData — Type

BlockData{T, TV<:AbstractVector{T}, TX<:AbstractVector{TV}} <: AbstractVector{T}

A strictly ordered collection of AbstractVectors, representing a ragged array of data.

Very useful when working with GPPPs. For example

f = @gppp let
    f1 = GP(SEKernel())
    f2 = GP(Matern52Kernel())
    f3 = f1 + f2
end

# Specify a `BlockData` set that can be used to index into
# the `f2` and `f3` processes in `f`.
x = BlockData(
    GPPPInput(:f2, randn(4)),
    GPPPINput(:f3, randn(3)),
)

# Index into `f` at the input.
f(x)

Base.split — Function

Base.split(x::BlockData, Y::AbstractVecOrMat)

Convenience functionality to make working with the output of operations on GPPPs easier. Splits up the rows of Y according to the sizes of the data in x.

julia> Y = vcat(randn(5, 3), randn(4, 3));

julia> x = BlockData(randn(5), randn(4));

julia> Y1, Y2 = split(x, Y);

julia> Y1 == Y[1:5, :]
true

julia> Y2 == Y[6:end, :]
true

Works with any BlockData, so blocks can e.g. be GPPPInputs. This is particularly helpful for working with the output from rand and marginals from a GPPP indexed at BlockData. For example

f = @gppp let
    f1 = GP(SEKernel())
    f2 = GP(Matern52Kernel())
    f3 = f1 + f2
end

x = BlockData(GPPPInput(:f2, randn(3)), GPPPInput(:f3, randn(4)))
y = rand(f(x))
y2, y3 = split(x, y)

Functionality also works with any AbstractVector.

Building GPPPs

If in doubt about what any of these transformations can do, play around with them!

Base.:+ — Function

+(fa::AbstractGP, fb::AbstractGP)

Produces an AbstractGP f satisfying f(x) = fa(x) + fb(x).

Base.:* — Function

*(f, g::AbstractGP)

Produce an AbstractGPh satisfying to h(x) = f(x) * g(x), for some deterministic function f.

If f isa Real, then h(x) = f * g(x).

Stheno.stretch — Function

stretch(f::AbstractGP, l::Union{AbstractVecOrMat{<:Real}, Real})

This is the primary mechanism by which to introduce length scales to your programme.

If l isa Real or l isa AbstractMatrix{<:Real}, stretch(f, l)(x) == f(l * x) for any input x. In the l isa Real case, this is equivalent to scaling the length scale by 1 / l.

l isa AbstractVector{<:Real} is equivalent to stretch(f, Diagonal(l)).

Equivalent to f ∘ Stretch(l).

Stheno.periodic — Function

periodic(g::AbstractGP, f::Real)

Produce an AbstractGP with period f.

Stheno.shift — Function

shift(f::AbstractGP, a::Real)
shift(f::AbstractGP, a::AbstractVector{<:Real})

Returns the GP g given by g(x) = f(x - a)

Stheno.select — Function

select(f::AbstractGP, idx)

Select the dimensions of the input to f given by idx.