GP API

GP API

This documents the user-facing API as it relates to the GP object.

This is a more thorough introduction to the internals than the Getting Started guide, which should be refered to if you are new to Stheno.jl. It's somewhere between a reference document and a tutorial.

GP

The primitive GP type is one of the core components of Stheno.jl. A GP should be thought of as a distribution over real-valued functions, in the same way that a Distributions.Normal is a distribution over real numbers, and Distibutions.MvNormal is a distribution over real-valued vectors.

Stheno.GP — Type.

GP{Tm<:MeanFunction, Tk<:Kernel}

A Gaussian Process (GP) with known mean m and kernel k. A book-keeping object gpc is also required, but only matters when composing GPs together.

Zero Mean

If only two arguments are provided, assume the mean to be zero everywhere:

julia> f = GP(Matern32(), GPC());

julia> x = randn(5);

julia> mean(f(x)) == zeros(5)
true

julia> cov(f(x)) == Stheno.pw(Matern32(), x)
true

Constant Mean

If a Real is provided as the first argument, assume the mean function is constant with that value

julia> f = GP(5.0, EQ(), GPC());

julia> x = randn(5);

julia> mean(f(x)) == 5.0 .* ones(5)
true

julia> cov(f(x)) == Stheno.pw(EQ(), x)
true

Custom Mean

Provide an arbitrary function to compute the mean:

julia> f = GP(x -> sin(x) + cos(x / 2), RQ(3.2), GPC());

julia> x = randn(5);

julia> mean(f(x)) == sin.(x) .+ cos.(x ./ 2)
true

julia> cov(f(x)) == Stheno.pw(RQ(3.2), x)
true

source

FiniteGP

Once constructed, the correct way to interact with a GP is via a FiniteGP, which is just the multivariate Normal given by considering the GP at only a finite set of inputs.

f = GP(Matern52(), GPC())
x = randn(10)
fx = f(x)

here fx, to be read as "f at x", is a FiniteGP to which the following methods apply:

Statistics.mean — Method.

mean(fx::FiniteGP)

Compute the mean vector of fx.

julia> f = GP(Matern52(), GPC());

julia> x = randn(11);

julia> mean(f(x)) == zeros(11)
true

source

Statistics.cov — Method.

cov(f::FiniteGP)

Compute the covariance matrix of fx.

Noise-free observations

julia> f = GP(Matern52(), GPC());

julia> x = randn(11);

julia> # Noise-free

julia> cov(f(x)) == Stheno.pw(Matern52(), x)
true

Isotropic observation noise

julia> cov(f(x, 0.1)) == Stheno.pw(Matern52(), x) + 0.1 * I
true

Independent anisotropic observation noise

julia> s = rand(11);

julia> cov(f(x, s)) == Stheno.pw(Matern52(), x) + Diagonal(s)
true

Correlated observation noise

julia> A = randn(11, 11); S = A'A;

julia> cov(f(x, S)) == Stheno.pw(Matern52(), x) + S
true

source

Statistics.cov — Method.

cov(fx::FiniteGP, gx::FiniteGP)

Compute the cross-covariance matrix between fx and gx.

julia> f = GP(Matern32(), GPC());

julia> x1 = randn(11);

julia> x2 = randn(13);

julia> cov(f(x1), f(x2)) == pw(Matern32(), x1, x2)
true

source

Stheno.marginals — Method.

marginals(f::FiniteGP)

Compute a vector of Normal distributions representing the marginals of f efficiently. In particular, the off-diagonal elements of cov(f(x)) are never computed.

julia> f = GP(Matern32(), GPC());

julia> x = randn(11);

julia> fs = marginals(f(x));

julia> mean.(fs) == mean(f(x))
true

julia> std.(fs) == sqrt.(diag(cov(f(x))))
true

source

Base.rand — Method.

rand(rng::AbstractRNG, f::FiniteGP, N::Int=1)

Obtain N independent samples from the marginals f using rng. Single-sample methods produce a length(f) vector. Multi-sample methods produce a length(f) x NMatrix.

julia> f = GP(Matern32(), GPC());

julia> x = randn(11);

julia> rand(f(x)) isa Vector{Float64}
true

julia> rand(MersenneTwister(123456), f(x)) isa Vector{Float64}
true

julia> rand(f(x), 3) isa Matrix{Float64}
true

julia> rand(MersenneTwister(123456), f(x), 3) isa Matrix{Float64}
true

source

Distributions.logpdf — Method.

logpdf(f::FiniteGP, y::AbstractVecOrMat{<:Real})

The logpdf of y under f if is y isa AbstractVector. logpdf of each column of y if y isa Matrix.

julia> f = GP(Matern32(), GPC());

julia> x = randn(11);

julia> y = rand(f(x));

julia> logpdf(f(x), y) isa Real
true

julia> Y = rand(f(x), 3);

julia> logpdf(f(x), Y) isa AbstractVector{<:Real}
true

source

Stheno.elbo — Method.

elbo(f::FiniteGP, y::AbstractVector{<:Real}, u::FiniteGP)

The saturated Titsias Evidence LOwer Bound (ELBO) [1]. y are observations of f, and u are pseudo-points.

julia> f = GP(Matern52(), GPC());

julia> x = randn(1000);

julia> z = range(-5.0, 5.0; length=13);

julia> y = rand(f(x, 0.1));

julia> elbo(f(x, 0.1), y, f(z)) < logpdf(f(x, 0.1), y)
true

[1] - M. K. Titsias. "Variational learning of inducing variables in sparse Gaussian processes". In: Proceedings of the Twelfth International Conference on Artificial Intelligence and Statistics. 2009.

source

Kernels

This is an ever-growing list. Implementing another kernel would make an excellent first PR...

Stheno.EQ — Type.

EQ() <: BaseKernel

The standardised Exponentiated Quadratic kernel. a.k.a. the Radial Basis Function (RBF), or Squared Exponential kernel.

$k(x, x^\prime) = \exp( -\frac{1}{2} || x - x^\prime||_2^2 )$

For length scales etc see stretch, for variance see *.

source

Stheno.Matern12 — Type.

Matern12 <: BaseKernel

The standardised Matern-1/2 / Exponential kernel:

$k(x, x^\prime) = \exp(-||x - x^\prime||_2)$

For length scales etc see stretch, for variance see *.

source

Stheno.Exp — Type.

Exp <: Kernel

The standardised Exponential kernel. Equivalent to Matern12.

For length scales etc see stretch, for variance see *.

source

Stheno.Matern32 — Type.

Matern32 <: BaseKernel

The standardised Matern kernel with ν = 3 / 2.

For length scales etc see stretch, for variance see *.

source

Stheno.Matern52 — Type.

Matern52 <: BaseKernel

The standardised Matern kernel with ν = 5 / 2.

For length scales etc see stretch, for variance see *.

source

Stheno.RQ — Type.

RQ <: BaseKernel

The standardised Rational Quadratic, with kurtosis α.

For length scales etc see stretch, for variance see *.

source

Stheno.Cosine — Type.

Cosine <: BaseKernel

Cosine BaseKernel with period parameter p.

source

Stheno.Linear — Type.

Linear{T<:Real} <: BaseKernel

The standardised linear kernel / dot-product kernel.

source

Stheno.Poly — Type.

Poly{Tσ<:Real} <: BaseKernel

Inhomogeneous Polynomial kernel. Poly(p, σ²) creates a Poly{p} with variance σ², defined as

k(xl, xr) = (dot(xl, xr) + σ²)^p

source

Stheno.GammaExp — Type.

GammaExp <: BaseKernel

The γ-Exponential kernel, 0 < γ ⩽ 2, is given by k(xl, xr) = exp(-||xl - xr||^γ).

source

Stheno.Wiener — Type.

Wiener <: BaseKernel

The standardised stationary Wiener-process kernel.

source

Stheno.WienerVelocity — Type.

WienerVelocity <: BaseKernel

The standardised WienerVelocity kernel.

source

Transformations of Kernels

Missing docstring.

Missing docstring for stretch(::Stheno.Kernel, ::Union{Real, AbstractVecOrMat{<:Real}}). Check Documenter's build log for details.

Base.:* — Method.

*(σ²::Real, k::Kernel)
*(k::Kernel, σ²::Real)

The right way to choose the variance of a kernel. Specifically, construct a kernel that scales the output of k by σ²:

julia> k = EQ();

julia> x = randn(11);

julia> pw(0.5 * k, x) == 0.5 .* Stheno.pw(k, x)
true

source

Base.:+ — Method.

+(kl::Kernel, kr::Kernel)

Construct the kernel whose value is given by the sum of those of kl and kr.

julia> kl, kr = EQ(), Matern32();

julia> x = randn(11);

julia> pw(kl + kr, x) == pw(kl, x) + pw(kr, x)
true

source

Base.:* — Method.

*(kl::Kernel, kr::Kernel)

Construct the kernel whose value is given by the product of those of kl and kr.

julia> kl, kr = EQ(), Matern32();

julia> x = randn(11);

julia> pw(kl * kr, x) == pw(kl, x) .* pw(kr, x)
true

source

Kernel Convenience

Stheno.kernel — Function.

kernel(k::Kernel;  l::Real=nothing, s::Real=nothing)

Convenience functionality to provide a kernel with a length scale l, and to scale the variance of k by s. Simply applies the stretch and * functions.

julia> k1 = kernel(EQ(); l=1.1, s=0.9);

julia> k2 = 0.9 * stretch(EQ(), 1 / 1.1);

julia> x = randn(11);

julia> pw(k1, x) == pw(k2, x)
true

source

MeanFunctions

These are implicit. Please refer to the GP documentation for details.